Answered Question 4 lab 2`
finished answering question 4 in the lab. You can add it to the file we are going to submit when you can. I was thinking of having predefined options for the "payment_plan" field but the command I found did not work for me. I will leave it here in case you want to try it:

ALTER TABLE SIGN_IN_INFO ALTER COLUMN payment_plan VARCHAR(30) NOT NULL CHECK (payment_plan IN ('Traditional', 'Blended', 'Online', 'Crash Course'));